requirements.txt
python-telegram-bot==22.5
tweepy==4.16.0
